+++ lib/eve_online/esi/character_industry_jobs.rb
@@ -1,7 +1,7 @@
 module EveOnline
   module ESI
     class CharacterIndustryJobs < Base
-      API_ENDPOINT = 'https://esi.tech.ccp.is/v1/characters/%s/industry/jobs/?datasource=tranquility&include_completed=%s'.freeze
+      API_ENDPOINT = 'https://esi.tech.ccp.is/v1/characters/%<character_id>s/industry/jobs/?datasource=tranquility&include_completed=%<include_completed>s'.freeze
 
       attr_reader :character_id, :include_completed
 
@@ -26,7 +26,7 @@ module EveOnline
       end
 
       def url
-        API_ENDPOINT % [character_id, include_completed]
+        format(API_ENDPOINT, character_id: character_id, include_completed: include_completed)
       end
     end
   end